+++ api/validators.go
@@ -5,34 +5,62 @@ import (
 	"fmt"
 	"github.com/MinterTeam/minter-go-node/core/state"
 	"github.com/MinterTeam/minter-go-node/core/types"
-	"github.com/MinterTeam/minter-go-node/core/validators"
 	"net/http"
 	"strconv"
 )
+
+type Stake struct {
+	Owner    types.Address    `json:"owner"`
+	Coin     types.CoinSymbol `json:"coin"`
+	Value    string           `json:"value"`
+	BipValue string           `json:"bip_value"`
+}
 
 type Candidate struct {
 	CandidateAddress types.Address `json:"candidate_address"`
 	TotalStake       string        `json:"total_stake"`
 	PubKey           string        `json:"pub_key"`
 	Commission       uint          `json:"commission"`
-	AccumReward      string        `json:"accumulated_reward"`
-	Stakes           []state.Stake `json:"stakes"`
+	Stakes           []Stake       `json:"stakes"`
 	CreatedAtBlock   uint          `json:"created_at_block"`
 	Status           byte          `json:"status"`
-	AbsentTimes      uint          `json:"absent_times"`
 }
 
-func makeResponseCandidate(c state.Candidate) Candidate {
+type Validator struct {
+	AccumReward string    `json:"accumulated_reward"`
+	AbsentTimes uint      `json:"absent_times"`
+	Candidate   Candidate `json:"candidate"`
+}
+
+func makeResponseValidator(v state.Validator, state *state.StateDB) Validator {
+	return Validator{
+		AccumReward: v.AccumReward.String(),
+		AbsentTimes: v.AbsentTimes,
+		Candidate:   makeResponseCandidate(*state.GetStateCandidate(v.PubKey), state),
+	}
+}
+
+func makeResponseCandidate(c state.Candidate, state *state.StateDB) Candidate {
+
+	stakes := make([]Stake, len(c.Stakes))
+
+	for i, stake := range c.Stakes {
+		stakes[i] = Stake{
+			Owner:    stake.Owner,
+			Coin:     stake.Coin,
+			Value:    stake.Value.String(),
+			BipValue: stake.BipValue(state).String(),
+		}
+	}
+
 	return Candidate{
 		CandidateAddress: c.CandidateAddress,
 		TotalStake:       c.TotalBipStake.String(),
 		PubKey:           fmt.Sprintf("Mp%x", c.PubKey),
 		Commission:       c.Commission,
-		AccumReward:      c.AccumReward.String(),
-		Stakes:           c.Stakes,
+		Stakes:           stakes,
 		CreatedAtBlock:   c.CreatedAtBlock,
 		Status:           c.Status,
-		AbsentTimes:      c.AbsentTimes,
 	}
 }
 
@@ -44,19 +72,20 @@ func GetValidators(w http.ResponseWriter, r *http.Request) {
 		height = int(blockchain.Height())
 	}
 
-	_, candidates := GetStateForRequest(r).GetValidators(validators.GetValidatorsCountForBlock(uint64(height)))
+	rState := GetStateForRequest(r)
+	vals := rState.GetStateValidators().Data()
 
 	w.Header().Set("Content-Type", "application/json; charset=UTF-8")
 	w.WriteHeader(http.StatusOK)
 
-	var responseCandidates []Candidate
+	var responseValidators []Validator
 
-	for _, candidate := range candidates {
-		responseCandidates = append(responseCandidates, makeResponseCandidate(candidate))
+	for _, val := range vals {
+		responseValidators = append(responseValidators, makeResponseValidator(val, rState))
 	}
 
 	json.NewEncoder(w).Encode(Response{
 		Code:   0,
-		Result: responseCandidates,
+		Result: responseValidators,
 	})
 }