+++ core/dao/dao.go
@@ -3,6 +3,6 @@ package dao
 import "github.com/MinterTeam/minter-go-node/core/types"
 
 var (
-	Address    = types.HexToAddress("Mxa93163fdf10724dc4785ff5cbfb9ac0b5949409f")
-	Commission = 5
+	Address    = types.HexToAddress("Mxee81347211c72524338f9680072af90744333146")
+	Commission = 10
 )