+++ core/developers/developers.go
@@ -1,0 +1,8 @@
+package developers
+
+import "github.com/MinterTeam/minter-go-node/core/types"
+
+var (
+	Address    = types.HexToAddress("Mx444c4f1953ea170f74eabef4eee52ed8276a7d5e")
+	Commission = 10
+)