+++ genesis/genesis.go
@@ -31,18 +31,18 @@ func GetTestnetGenesis() (*tmtypes.GenesisDoc, error) {
 	}
 
 	appState := AppState{
-		FirstValidatorAddress: types.HexToAddress("Mxa93163fdf10724dc4785ff5cbfb9ac0b5949409f"),
+		FirstValidatorAddress: types.HexToAddress("Mxee81347211c72524338f9680072af90744333146"),
 		InitialBalances: []Account{
 			{
-				Address: types.HexToAddress("Mxa93163fdf10724dc4785ff5cbfb9ac0b5949409f"),
+				Address: types.HexToAddress("Mxee81347211c72524338f9680072af90744333146"),
 				Balance: map[string]string{
-					"MNT": helpers.BipToPip(big.NewInt(1000000000)).String(),
+					"MNT": helpers.BipToPip(big.NewInt(100000000)).String(),
 				},
 			},
 			{
 				Address: types.HexToAddress("Mxfe60014a6e9ac91618f5d1cab3fd58cded61ee99"),
 				Balance: map[string]string{
-					"MNT": helpers.BipToPip(big.NewInt(10000000)).String(),
+					"MNT": helpers.BipToPip(big.NewInt(100000000)).String(),
 				},
 			},
 		},
@@ -55,8 +55,8 @@ func GetTestnetGenesis() (*tmtypes.GenesisDoc, error) {
 	}
 
 	genesis := tmtypes.GenesisDoc{
-		GenesisTime:     time.Date(2018, 7, 31, 0, 0, 0, 0, time.UTC),
-		ChainID:         "minter-test-network-17",
+		ChainID:         "minter-test-network-18",
+		GenesisTime:     time.Date(2018, 8, 22, 0, 0, 0, 0, time.UTC),
 		ConsensusParams: nil,
 		Validators: []tmtypes.GenesisValidator{
 			{